Move message brokers to the first level of the package
