+++ tests/test_integration/test_message_brokers.py
@@ -2,7 +2,7 @@ import json
 
 import redis.asyncio as redis
 
-from diator.events.message_brokers.redis import Message, RedisMessageBroker
+from diator.message_brokers.redis import Message, RedisMessageBroker
 
 
 async def test_redis_message_broker_publish_event(
