+++ tests/test_unit/test_request_response.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass, field
 from uuid import UUID, uuid4
 
 from diator.events import Event, EventEmitter, EventMap
-from diator.events.message_brokers.redis import Message
 from diator.mediator import Mediator
+from diator.message_brokers.redis import Message
 from diator.requests import Request, RequestHandler, RequestMap
 from diator.response import Response
